pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "configguard"
version = "1.0.0"
description = "AI-driven network configuration compliance platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Corey A. Wade", email = "corey@coreywade.com"},
]
keywords = ["network", "compliance", "security", "NIST", "CIS", "PCI-DSS", "configuration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "flask>=3.0",
    "rich>=13.0",
    "reportlab>=4.0",
    "pydantic>=2.0",
    "watchdog>=3.0",
    "gitpython>=3.1",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "httpx>=0.25",
]
batfish = [
    "pybatfish>=2023.10",
]

[project.scripts]
configguard = "configguard.cli:cli"

[project.urls]
Homepage = "https://github.com/cwccie/configguard"
Documentation = "https://github.com/cwccie/configguard#readme"
Repository = "https://github.com/cwccie/configguard"
Issues = "https://github.com/cwccie/configguard/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
configguard = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
